+++ mods/src/input/input.cpp
@@ -23,7 +23,7 @@ void init_input() {
     _init_attack();
 
     // Disable Raw Mouse Motion
-    if (feature_has("Disable Raw Mouse Motion (Not Recommended)", server_disabled)) {
+    if (feature_has("Disable Raw Mouse Motion", server_disabled)) {
         media_set_raw_mouse_motion_enabled(false);
     }